requirements.txt
discord.py
youtube-dl
requests
ffmpeg-python
python-dotenv
PyNaCl
validators